+++ src/meta.ts
@@ -1,2 +1,2 @@
 export const PACKAGE_NAME = 'tia-gateway'
-export const PACKAGE_VERSION = '0.2.0'
+export const PACKAGE_VERSION = '0.4.0'